Renamed the store phone locations script and changed the getPrices script to index to allow for CI.
