+++ index.js
@@ -1,95 +1,85 @@
-// The LambdaLocationUpdateHandler is responsible for responding to requests from clients to
-// update their current location entry in the database.
-//
-// The DynamoDB locations table can be viewed here:
-// https://eu-west-2.console.aws.amazon.com/dynamodb/home?region=eu-west-2#tables:selected=locations
-//
-// This code must be copied over to the AWS Lambda Management Console here:
-// https://eu-west-2.console.aws.amazon.com/lambda/home?region=eu-west-2#/functions/HandleLocationUpdate
+exports.handler         = (event, context, callback) => {
+    // Setup
+    var AWS             = require("aws-sdk");
+    var docClient       = new AWS.DynamoDB.DocumentClient();
+ 
+    //Carry out input validation 
+    let validationVar   = eventValidation(event);
 
-var doc = require('dynamodb-doc');
-var dynamodb = new doc.DynamoDB();
 
-// Name of the locations table in DynamoDB.
-let k_TABLE_NAME = "locations";
-
-// This function is called in response to a request from a client.
-//
-// The `event` parameter holds information about the request, including the parameters.
-// The following parameters are expected in the request:
-//  - `userId` a unique identifier (e.g. email address) for this device (used as DB primary key).
-//  - `latitude` the current latitude of the user (expected in ISO 6709 format - https://en.wikipedia.org/wiki/ISO_6709).
-//  - `longitude` the current longitude of the user (expected in ISO 6709 format.)
-
-// To print a log in the console use -> console.log('value1 =', event.key1);
-
-exports.handler = (event, context, callback) => {
-    // Carry out input validation on the request's parameters.
-    let validationVar = eventValidation(event);
+    // User Inputs returned from validation. 
     if(validationVar[0]){
-        let userIdInput = validationVar[1];
-        let latitudeInput = validationVar[2];
-        let longitudeInput = validationVar[3];
+        let distance    = validationVar[1];
+        let latitude    = validationVar[2];
+        let longitude   = validationVar[3];
     } else {
         abortLocationUpdate("Error with input validation", callback);
+    }    
+
+    // Works out the radius of LAT/LON values to be accepted. 
+    var start_lat       = (latitude - (1/(distance/(110.574*1000))));
+    var end_lat         = (latitude + (1/(distance/(110.574*1000))));
+    var start_lon       = (longitude - (1/(distance/(111.320*Math.cos(start_lat)))));
+    var end_lon         = (longitude + (1/(distance/(111.320*Math.cos(end_lat)))));
+    
+    // Retrieves items from the DB. 
+    var params = getParams(start_lat, end_lat, start_lon, end_lon);
+ 
+    //List of items that will be returned. Needed to account for DynamoDB window size. 
+    var items = []
+    
+    // Accounts for DynamoDB only being able to send a set window at a time, scans whole DB. 
+    var scanExecute = function(callback) {
+     
+        docClient.scan(params,function(err,result) {
+ 
+            if(err) {
+                callback(err);
+            } else {
+                 
+                items = items.concat(result.Items);
+ 
+                if(result.LastEvaluatedKey) {
+ 
+                    params.ExclusiveStartKey = result.LastEvaluatedKey;
+                    scanExecute(callback);              
+                } else {
+                    
+                    // This is where the final items are returned. 
+                    callback(err,items);      // **** Uncomment this and comment all below if you just want all prices returned. 
+                    
+
+                    // // Add in the averaging functionality here. 
+                    // var latitudeRange = Math.range(items.latitude);
+                    // console.print("LAT RANGE:   " + latitudeRange);
+                    
+                    // var sum = 0;
+                    // for(var i = 0, l = items.length; i < l; i++){
+                    //      sum += parseInt( items[i].price, 10 ); //don't forget to add the base
+                    // }
+                    // var avg = sum/items.length;
+                    
+                    // callback(err,avg);
+                }   
+            }
+        });
     }
-    // At this point, we assume that the input is valid and correctly formed.
-
-    // Make a note of the current time.
-    let currentDateTime = new Date().toISOString();
-
-    // Create the new location entry item.
-    var locationItem = {
-        userId      : userIdInput,
-        date        : currentDateTime,
-        latitude    : latitudeInput,
-        longitude   : longitudeInput
-    };
-
-    // Insert a new record into the `locations` DynamoDB table.
-    dynamodb.putItem({
-        "TableName" : k_TABLE_NAME,
-        "Item" : locationItem
-    }, function(error, data) {
-        if (error) {
-            abortLocationUpdate("Failed inserting location update into DynamoDB: " + error, callback);
-        } else {
-            callback(null, {
-                "statusCode" : 200,
-                "headers" : { "Content-Type" : "application/json" },
-                "body" : JSON.stringify({
-                    "status" : "success",
-                    "message" : "Successfully updated location"
-                })
-            });
-        }
-    });
+    
+    scanExecute(callback);
 };
-
-// This function simply reports an error back to the client.
-function abortLocationUpdate(reason, callback) {
-    callback(null, {
-        "statusCode" : 200,
-        "headers" : { "Content-Type" : "application/json" },
-        "body" : JSON.stringify({
-            "status" : "error",
-            "message" : reason
-        })
-    });
-}
 
 // This is a function to perform input validation on the inputs in event.
 function eventValidation(e){
   
     let event = e;
-    let userIdInput = event.queryStringParameters.userId;
+    let userDistance = event.queryStringParameters.distance;
     let latitudeInput = event.queryStringParameters.latitude;
     let longitudeInput = event.queryStringParameters.longitude;
 
-    if (!(userIdInput === null || latitudeInput === null || longitudeInput === null)) {
+    if (!(userDistance === null || latitudeInput === null || longitudeInput === null)) {
         // Extract the userId parameter and validate.  
-        if (!(hexReg(userIdInput))){
-            abortLocationUpdate("Invalid user ID", callback);
+        if (Number.isInteger(Distance)){
+            abortLocationUpdate("Distance not a Number", callback);
         }
         
         // Extract the `latitude` parameter and validate.       
@@ -103,22 +93,33 @@ function eventValidation(e){
         }
 
         // If here the validation was a success *Martin Cheers*
-        return [true, userIdInput, latitudeInput, longitudeInput]
+        return [true, userDistance, latitudeInput, longitudeInput]
 
     } else {
         abortLocationUpdate("Null value at input", callback);
     }
 }
 
-// this is a function to check for a hexidecimal value and a length of 64 characters.
-function hexReg(s) {
-    var regExp = /[0-9A-Fa-f]{16}/g;
-    return (regExp.test(s));
-}
-
-// this function checks the latitude and lonitude follow the correct format.
-function longLatReg(l){
-    // regex for latitude and longitude.
-    var regExp = /(\-?\d+(\.\d+)?)/;
-    return regExp.test(l);
+function getParams(start_lat, end_lat, start_lon, end_lon){
+    // Retrieves items from the DB. 
+    var params = {
+        TableName:"price_paid_data",
+        
+        // What fields will be returned. 
+        ProjectionExpression: "latitude , longitude, price",
+        
+        // Filter the items to only include LAT/LON within the radius. 
+        FilterExpression: "#latitude BETWEEN :latFrom AND :latTo and #longitude BETWEEN :lonFrom AND :lonTo",
+        ExpressionAttributeNames: {
+            "#latitude": "latitude",
+            "#longitude": "longitude"
+        },
+        ExpressionAttributeValues: {
+             ":latFrom" :   Math.min(start_lat,end_lat), //Math.max / Math.min to account for negatice numbers and avoid an error with the between comparison. 
+             ":latTo"   :   Math.max(start_lat,end_lat), 
+             ":lonFrom" :   Math.min(start_lon, end_lon),
+             ":lonTo"   :   Math.max(start_lon, end_lon)
+        }
+    };
+    return params
 }
